Locate cbc system library with pkg-config
This adds a tiny build-time dependency on pkg-config. Note that if
`pkg-config` is not present, or is present but fails to locate the
library or execute, nothing weird happens: the user gets the same
behaviour they've had until now.

I've added a line to README to mention `pkg-config` too.

Fixes #23